[Refactor] 메인페이지 검색 분리
