+++ src/components/SearchBar.tsx
@@ -5,7 +5,7 @@ import RadioGroup from './common/RadioGroup'
 import { useState } from 'react'
 import CareerSelect from './common/CareerSelect'
 import { useDebounce } from '../hooks/useDebounce'
-import { type SearchParams } from '../pages/Home'
+import type { SearchParams } from '@/hooks/useSearchPortfoilo'
 
 const INTEREST_SELECT = [
   { label: '전체', value: 'all' },
@@ -21,10 +21,10 @@ const INTEREST_SELECT = [
 ]
 
 interface SearchBarProps {
-    onSearch: (params: SearchParams) => void
-  }
-  
-export const SearchBar = ({ onSearch }:SearchBarProps) => {
+  onSearch: (params: SearchParams) => void
+}
+
+export const SearchBar = ({ onSearch }: SearchBarProps) => {
   const [interest, setInterest] = useState<string>('all')
   const [career, setCareer] = useState('')
   const [keyword, setKeyword] = useState('')
